refactor handlelinks
